# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Student Admission & Entry Test Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all pre-uploaded students
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: program
          schema:
            type: string
        - in: query
          name: eligible
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentListResponse"
    post:
      operationId: createStudent
      tags: [students]
      summary: Add a single student record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /students/{id}/block:
    patch:
      operationId: blockStudent
      tags: [students]
      summary: Block or unblock a suspicious student
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockStudentInput"
      responses:
        "200":
          description: Student block status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/bulk-import:
    post:
      operationId: bulkImportStudents
      tags: [students]
      summary: Bulk import students from CSV/Excel data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"

  # Verification
  /verification/check:
    post:
      operationId: checkStudentRecord
      tags: [verification]
      summary: Check roll number and CNIC before face verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationCheckInput"
      responses:
        "200":
          description: Record found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationCheckResponse"
        "404":
          description: No matching record
  /verification/attempt:
    post:
      operationId: submitVerificationAttempt
      tags: [verification]
      summary: Submit face verification attempt with captured image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationAttemptInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationAttemptResponse"
  /verification/logs:
    get:
      operationId: listVerificationLogs
      tags: [verification]
      summary: Admin - list all verification logs
      parameters:
        - in: query
          name: studentId
          schema:
            type: integer
        - in: query
          name: success
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Verification logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationLogListResponse"

  # Applications
...
  /tests:
    get:
      operationId: listTests
      tags: [tests]
      summary: List tests
      responses:
        "200":
          description: Tests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"
    post:
      operationId: createTest
      tags: [tests]
      summary: Create a test configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestInput"
      responses:
        "201":
          description: Test created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
...
  /tests/{id}:
    get:
      operationId: getTest
      tags: [tests]
      summary: Get test by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
    patch:
      operationId: updateTest
      tags: [tests]
      summary: Update test configuration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestUpdate"
      responses:
        "200":
          description: Updated test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
    delete:
      operationId: deleteTest
      tags: [tests]
      summary: Delete test
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Questions
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, rollNumber, cnic, name, fatherName, phone, program, isEligible, isBlocked, createdAt]
      properties:
        id:
          type: integer
        rollNumber:
          type: string
        cnic:
          type: string
        name:
          type: string
        fatherName:
          type: string
        phone:
          type: string
        schoolCollege:
          type: ["string", "null"]
        program:
          type: string
        photoUrl:
          type: ["string", "null"]
        isEligible:
          type: boolean
        isBlocked:
          type: boolean
        blockReason:
          type: ["string", "null"]
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [rollNumber, cnic, name, fatherName, phone, program]
      properties:
        rollNumber:
          type: string
        cnic:
          type: string
        name:
          type: string
        fatherName:
          type: string
        phone:
          type: string
        schoolCollege:
          type: string
        program:
          type: string
        photoUrl:
          type: string
        isEligible:
          type: boolean
    StudentUpdate:
      type: object
      properties:
        rollNumber:
          type: string
        cnic:
          type: string
        name:
          type: string
        fatherName:
          type: string
        phone:
          type: string
        schoolCollege:
          type: string
        program:
          type: string
        photoUrl:
          type: string
        isEligible:
          type: boolean
    StudentListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    BlockStudentInput:
      type: object
      required: [isBlocked]
      properties:
        isBlocked:
          type: boolean
        blockReason:
          type: string
    BulkImportInput:
      type: object
      required: [students]
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/StudentInput"
    BulkImportResult:
      type: object
      required: [imported, failed, total]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        total:
          type: integer
        errors:
          type: array
          items:
            type: string

    # Verification
    VerificationCheckInput:
      type: object
      required: [rollNumber, cnic]
      properties:
        rollNumber:
          type: string
        cnic:
          type: string
    VerificationCheckResponse:
      type: object
      required: [found, studentName, photoUrl]
      properties:
        found:
          type: boolean
        studentName:
          type: string
        photoUrl:
          type: string
        studentId:
          type: integer
    VerificationAttemptInput:
      type: object
      required: [studentId, capturedImageBase64, deviceInfo, ipAddress]
      properties:
        studentId:
          type: integer
        capturedImageBase64:
          type: string
        faceMatchScore:
          type: number
        deviceInfo:
          type: string
        ipAddress:
          type: string
        browserInfo:
          type: string
    VerificationAttemptResponse:
      type: object
      required: [success, message, sessionToken]
      properties:
        success:
          type: boolean
        message:
          type: string
        sessionToken:
          type: ["string", "null"]
        faceMatchScore:
          type: ["number", "null"]
        studentId:
          type: ["integer", "null"]
    VerificationLog:
      type: object
      required: [id, studentId, success, faceMatchScore, capturedImageUrl, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        success:
          type: boolean
        faceMatchScore:
          type: number
        capturedImageUrl:
          type: ["string", "null"]
        deviceInfo:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        browserInfo:
          type: ["string", "null"]
        createdAt:
          type: string
    VerificationLogListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/VerificationLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Application
...
    Test:
      type: object
      required: [id, title, durationMinutes, totalMarks, passingMarks, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        instructions:
          type: ["string", "null"]
        isActive:
          type: boolean
        questionCount:
          type: integer
        createdAt:
          type: string
    TestInput:
      type: object
      required: [title, durationMinutes, totalMarks, passingMarks]
      properties:
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        instructions:
          type: string
        isActive:
          type: boolean
    TestUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        instructions:
          type: string
        isActive:
          type: boolean

    # Question
...