# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Admission & Entry Test Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  rollNumber: string;
  cnic: string;
  name: string;
  fatherName: string;
  phone: string;
  schoolCollege?: string;
  program: string;
  photoUrl?: string;
  isEligible?: boolean;
}