# lib/api-zod/src/generated/types/verificationCheckResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Admission & Entry Test Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface VerificationCheckResponse {
  found: boolean;
  studentName: string;
  photoUrl: string;
  studentId?: number;
}